requirements.txt
click==8.1.3
colorama==0.4.6
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Flask==2.2.2
pymongo==4.1.1
bcrypt==4.0.1
Werkzeug==2.2.2
pycryptodome == 3.15.0
numpy == 1.26.4 